Definitely don't look at this code :/

+++ app/src/main/java/com/pxp200/krakenapp/model/User.java
@@ -21,7 +21,7 @@ public class User {
         buildings = new ArrayList<>();
     }
 
-    public void incrementResource(String resource, Integer cnt){
+    public void incrementResource(String resource, double cnt){
         //INCREMENT RESOURCE
 
     }
